refactor: modularize waterfall components and hooks
